refactor: Refactor CategoryController and CategoryService classes

+++ src/main/java/dev/JpDev17/movieflix/services/CategoryService.java
@@ -1,7 +1,11 @@
 package dev.JpDev17.movieflix.services;
 
 import dev.JpDev17.movieflix.entities.Category;
+import dev.JpDev17.movieflix.mapper.CategoryMapper;
 import dev.JpDev17.movieflix.repositories.CategoryRepository;
+import dev.JpDev17.movieflix.request.CategoryRequest;
+import dev.JpDev17.movieflix.response.CategoryResponse;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,19 +20,38 @@ public class CategoryService {
         this.categoryRepository = categoryRepository;
     }
 
-    public List<Category> getAll() {
-        return categoryRepository.findAll();
+    public List<CategoryResponse> getAll() {
+        List<Category> categoryList = categoryRepository.findAll();
+        return categoryList.stream()
+                .map(CategoryMapper::toCategoryResponse)
+                .toList();
     }
 
-    public Optional<Category> getById(Long id) {
-        return categoryRepository.findById(id);
+    public CategoryResponse getById(Long id) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Category with id " + id + " not found!"));
+        return CategoryMapper.toCategoryResponse(category);
     }
 
-    public Category createCategory(Category category) {
-        return categoryRepository.save(category);
+    public CategoryResponse createCategory(CategoryRequest request) {
+        Category cratedCategory = CategoryMapper.toCategory(request);
+        Category newCategory = categoryRepository.save(cratedCategory);
+        return CategoryMapper.toCategoryResponse(newCategory);
+    }
+    
+    public CategoryResponse updateCategory(Long id, CategoryRequest request) {
+        Category createdCategory = CategoryMapper.toCategory(request);
+        categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Category with id " + id + " not found!"));
+        createdCategory.setId(id);
+        Category newCategory = categoryRepository.save(createdCategory);
+        return CategoryMapper.toCategoryResponse(newCategory);
     }
 
     public void deleteById(Long id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new EntityNotFoundException("Category with id " + id + " not found!");
+        }
         categoryRepository.deleteById(id);
     }
 }
